Modifications to remove several common block and namelist files: contents of comgpsro.cdk, namgpsro.cdk and new s/r sugpsro are now in modgpsro_mod.ftn90; contents of comvfiles.cdk and setup_burpfiles.ftn90 are now in new module burpfiles_mod.ftn90. The s/r readnml.ftn was removed, since it is no longer used.
Also, replaced several conversion factors in bmatrixEnsemble_mod.ftn90 and bmatrixhi_mod.ftn90 with the use of parameters in mathPhysConstants_mod and earthConstants_mod.

NOTE: the changes to conversion factors (especially knots->m/s) changes the results during an EnVar analysis quite a bit - this will have to be tested in a short cycle! The new value for knots->m/s is now consistent with the one used in GEM, so this should be the "correct" value.

The results from a background check test were unchanged.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@380 3ecdeb99-2462-4cc2-86f5-4e70ad797af3