Added functionality to obsSpaceDiag_mod to compute innovation stddev and background-error (from total B matrix) stddev. These are output in ascii files for the requested families/elements. This calculation is skipped if the NAMOSD block is not present in the namelist.
Also uniformized all modules (other than the GPS cdk90 files) so that they contain a simple "private" command at the beginning making everything private by default, then explicitly specifying those variables/derived types/procedures that are public with a "public" command. Also moved some "use" commands from inside module procedures to the top of the module so that the dependencies between a given module and other modules is more easily visible by just looking at the beginning of each module. Renamed a few public module variables/procedures to add a module specific prefix. The goal of all these changes is to make the basic interface and dependencies of the modules easier to see.

Added a new module codtyp_mod.ftn90 that contains integer parameters with the codtyp values. So far only used in obsSpaceDiag_mod. Could be added to shared repository and used throughout the EnVar and EnKF code.

No impact on the results for both analysis and background check tests (not tested in LAM mode).



git-svn-id: svn://mrbsvn/assimilation/var/trunk@389 3ecdeb99-2462-4cc2-86f5-4e70ad797af3