WARNING!!! Major update for envar/3d-var analysis mode: now the analysis increment, B matrices and observations are distributed with respect to 2 dimensions for MPI (not backward compatible). Instead of using a topology of e.g. 1x80x4, you need to use something like 10x8x4. Because of this change, the results are slightly affected due to the change in order of summing the cost function. Currently, the program for splitting the ensemble members into files with a single lat-lon tile does not work correctly, so it is preferrable to not do this splitting and instead use horizontally global files (this results in very little impact on execution time since the MPI communication of the members was improved so that it is a lot faster than before).
Some changes were also made to correct problems when performing an ensemble of EnVar analyses. This results in increased overall memory useage at the time the analysis increment is written out and therefore it may be necessary to slightly increase the total number of cpus to increase to amount of available memory.

Minor changes to significantly speed up the initialization of the global NMC B matrix and the writing of the Hessian.

Cecilien is almost finished testing a maestro suite that validates this revision.

No impact for the background check, since it doesn't use a geographical decomposition of the observations for MPI.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@439 3ecdeb99-2462-4cc2-86f5-4e70ad797af3