Major modification for ensemble-EnVar: perturbations are now added to the analysis increments for the perturbed members using the Bnmc matrix and scaling factors specified in the namelist (default values are .66 and .33 for e1 and e2). Three sets of increment files are now output when in ensemble-EnVar mode (rebm0: no perturbation added, rebm1: perturbation with "e1" factor, rebm2: perturbation with "e2" factor).
Also, rpn_comm_allreduce has been replaced with mpi_allreduce_sumreal8scalar (in mpi_mod.ftn90) for all cases when summing a real*8 scalar (mostly what is used in EnVar). This eliminates small differences in the results that occured previously when changing the number of openMP threads (results still differ when changing MPI topology).

Reduced the amount of stuff printed in the listing from the setup of the global spectral transform and the gaussian grid modules.

Results in analysis mode are slightly changed due to the change in the order of summing the cost function from replacing rpn_comm_allreduce.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@442 3ecdeb99-2462-4cc2-86f5-4e70ad797af3