Removed/changed dependencies between modules so that the enkf_pturb program now requires a lot less code for compiling. Also modified this program so that it again works correctly with latest changes.
Move subasic_gd.ftn into writeincrement_mod.ftn90 to take advantage of explicit interface by being in a module. Other modifications to writeincrement_mod.ftn90 so that if the user does not request GZ or VT or ES in the increment, then the TT background state is not read in and these variables are not calculated.

Modified minimization_mod.ftn90 to allow increment from control analysis to be used as the first guess for the perturbed analyses. Also, can now use Hessian from control analysis for perturbed analyses. Both of these features seem helpful for speeding up convergence of perturbed members.

Added an extra check oin burp_read_mod.ftn90 to handle case found with split burp files and no observations in a file. Not sure why this was needed. Possibly only because the program for splitting the burp files was not working correctly for more than about 90 mpi processes.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@444 3ecdeb99-2462-4cc2-86f5-4e70ad797af3