Minor modifications to fix bugs when using burp files already split by lat-lon tiles. This was not tested since the change of mpi decomposition from latitude bands to lat-lon tiles.
Minor bug fixed in bmatrixensemble of uninitialized array that caused a floating point error only when testing with 512 mpi tasks.

Also, reintroduced the non-square formulation of the B matrix sqrt, but now controlled with the namelist logical variable "SquareSqrt" so that both formulations can be used (default is false, i.e. the old non-square formulation). The old version of dsyev was reintroduced for backward compatibility when using old formulation, but rename dsyev2 and put in the main trunk directory (previously in modulopt).

For ens-envar, removed the namelist variable "lxbar_fg_only" and replaced with "lvazx" which is an old namelist variable that enables an existing mechanism for reading the increment stored in the pm1q file as a starting point for the minimization. Also, added new namelist variable NITERMAX_PERT to allow specifying a different number of iterations for the perturbed analyses than for the control analysis.

NOTE: WARNING: because of the change in the default behaviour regarding the calculation of the sqrt of the static B matrix, a change is required in the namelist to obtain the same behaviour as the previous version. The default behaviour is again like it was in version 405 and before. Otherwise results should be unchanged.





git-svn-id: svn://mrbsvn/assimilation/var/trunk@450 3ecdeb99-2462-4cc2-86f5-4e70ad797af3