Two optional tweaks to improve the representation of TG background error covariances are introduced.
In B_HI (global mode only), the legacy TG std. dev. from ./bgcov can be adapted based on the land-sea and the sea-ice masks from the trial field to make sure that realistic values are used over land/ice and open water. Users can activate this option using:

&NAMBHI
 TWEAKTG = .TRUE.

In B_ENS, to alleviate the absence of spread for TG in the EnKF (because the SST is not perturbed), it is possible to replace, only over open water, the TG ensemble perturbations by the 2m T ensemble perturbations. Users can activate this option using:

&NAMBEN
 TWEAKTG = .TRUE.




git-svn-id: svn://mrbsvn/assimilation/var/trunk@497 3ecdeb99-2462-4cc2-86f5-4e70ad797af3