Modification of the following 2 routines/files:
  seterrgpsgb.ftn (sets GB-GPS ZTD observation errors)
  suinnov.ftn     (modified to pass background state (columnhr) to routine seterrgpsgb)

The modification to routine seterrgpsgb affects the assignment of ZTD observation errors for GPS 
sites that do not report surface pressure which is needed to set the ZTD obervation error. 

For cases where surface pressure is not reported, the routine now estimates surface pressure at 
the GPS site from the background state surface pressure and temperature. Before, a gross estimate
of surface pressure was made using the site elevation and an assumed sea-level pressure.

As this modification alters the ZTD observation errors assigned to some observations, it will not 
give the same exact results (analysis) as the previous EnVar revision.
--This line, and those below, will be ignored
M    suinnov.ftn
M    seterrgpsgb.ftn


git-svn-id: svn://mrbsvn/assimilation/var/trunk@615 3ecdeb99-2462-4cc2-86f5-4e70ad797af3