Modifications to compile and run with the Intel compiler on Linux. This code has also been tested on AIX-powerpc7.
Also updated the 'shared' directory to revision 141 with command:
    svn propset svn:externals 'shared -r141 svn://mrbsvn/assimilation/shared/trunk' .
which also contains modifications to run on Linux.

Unfortunatly, this code does not compile on Linux since the 'netcdf' library is not yet available on that platform.
We will find a solution later.

The rest of the code is compiling correctly on Linux.

This is equivalent to revision 630 in branch 'releases/v_2.0.a'.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@632 3ecdeb99-2462-4cc2-86f5-4e70ad797af3