This merges in the commit 649 from 'trunk'.
Modified bmatrixhi_mod.ftn90 so that it can read/write the sqrt of corns instead
of using the values computed in the code. Once the sqrt is stored in the stats file
this will remove the sensitivity to changes to the ESSL library.

Without any namelist change, this has no impact on results.

The command done was
    svn merge -c649 svn://mrbsvn/assimilation/var/trunk



git-svn-id: svn://mrbsvn/assimilation/var/releases/v_2.1@652 3ecdeb99-2462-4cc2-86f5-4e70ad797af3