Another semi-major modification. With these modifications the analysis increment grid in global mode now has latitudes ordered from south-to-north (i.e. Gaussian grid with ig2=0), making it consistent with how things are stored in RPN standard files and avoiding problems with the previous "flipped" Gaussian grid not being completely supported by the ezscint interpolation package.
NOTE: in global mode a different analysis grid prototype file must be used with this revision that specifies the Gaussian grid with ig2=0! These files can temporarily be found in the following directory: /users/dor/arma/gr3/data_cnfs/prototype_grid/nonflipped/

NOTE: The results for an analysis are slightly changed, due to a difference in the order of summation in the cost function. Also, most likely the Hessian preconditioning file obtained with previous revisions may not be optimal when used with this revision.

The program calcstats was changed for computing global NMC background-error statistics, but not yet tested.



git-svn-id: svn://mrbsvn/assimilation/var/trunk@701 3ecdeb99-2462-4cc2-86f5-4e70ad797af3