Changed order of indices in glb spectral transform
Consistent with previous changes to gridstatevector_mod.ftn90, the order of indices used in the global spectral transform and throughout the B matrix modules is now (lon,lon,lev). This was also done throughout the code for the global B matrix calculation (in ./calcstats/). Some likely bugs (related to the changes to gridstatevector_mod.ftn90 in the previous commit) in the background check code were also fixed, but not yet tested. Some minor changes to the spectral transform should result in faster execution times (for the Legendre transform).

EnVar and global calcstats were both tested with no changes in results. Background check not tested.