Reconfigure the loops for ensemble B matrix
The loops were changed so that the contribution from each ensemble member index is computed (for all levels). Previously all members for a given level were computed. Related to this the control vector for the ensemble B is now distributed over mpi tasks with respect to spectral coefficient indices m and n, making it consistent with the NMC B matrix. Because of this the global spectral transform was simplified to eliminate all other mpimodes.

The results are slightly changed by these modifications since the control vector for the ensemble B is distributed differently over mpi tasks. The code seems to be slightly faster in a realistic configuration.

These changes were only tested in global EnVar mode so far (need to test in LAM EnVar mode and for calcstats).