Extensive use of gridstatevector in bmatrixensemble
Major changes to bmatrixensemble_mod.ftn90 to make use of gridstatevector
objects for all gridded fields. This resulted in some code simplifications and
some improvements to efficiency. But the main motivation is to prepare for
scale-dependent localization with grids of varying resolution and also
a more efficient implementation of alpha advection/diffusion.

Gridstatevector was modified to be able to use either real(8) or real(4) fields.
This could later be generalized to a 16 bit integer or floating point
representation for saving more memory.