Added ability to advect ensemble amplitudes
Advection of ensemble amplitudes added to bmatrixensemble module. This is
control by new namelist variable "advectAmplitudeFactor" which gives the
multiplicative factor applied to the ensemble mean winds used for the
advection. The default value is 0.0 which bypasses all of the new
functionality.

This is mainly for use in FSOI calculations, since it is probably too
expensive for operational EnVar analysis, due to heavy mpi communication.

Also changed abort3d so that all mpi tasks are killed even if only one aborts.