New error trap in bmatrixEnsemble
Since the size of nkgdimEns is still hardwired while the size of the ensPert is now determined by NAMSTATE, a fatal error can occur if the appropriate variables are not set in NAMSTATE section of the namelist. This error trap force the code to abort if the size are not consistent.