Introducing 2DVar in LAM mode
It is now possible to generate 2D analysis using Bhi in LAM mode. To activate the new option, the following variable must be added to &NAMCT0 in analysis mode and to &NAMENS in calcstats mode.

SPATIALDIMENSIONS='2D',
IP1 = 93423264 (the ip1 of the 2D level,here hyb = 1.0)

This mode is introduced for future land surface analysis applications. The adaptation of Bens to allow 2DEnVar will follow in the near future.