Preliminary changes for vcode_anl=5005 and a bugfix
All changes required to perform a 3DVar analysis with vcode=5005
for the analysis increment have been made, except for changes possibly
needed to the ground-based GPS and GPS-RO linearized observation
operators.

During this work an important bug was fixed in gsv_readFromFile
which caused problems when no vertical interpolation was required
when reading in a file with multiple time levels. This would have
affected the increments of HU output in the rebm files when the
background state and analysis increment were on the same levels.