Added IBM compile script; documentation
For testing purposes, temporarily put back in a compile script that works for the IBM (spica/hadar). Also, started to standardize the in-code documentation for all files in the main source directory. This allows use of "f90doc" to automatically generate html documentation. Also add "_mod" to the few module names that did not already have it.