First version of oavar_system_tests suite
This maestro suite can be used to tests the oavar fortran code. The first test performs a small EnVar minimization that should run relatively quickly (only max of 1 node for all tasks). Also, small change to bmatrixensemble_mod so that in global mode it can read ensemble members either interpolated to the Gaussian grid or on the original Z grid (with the first longitude duplicated).