New namcodtyp useage; avoid tovs calls
Now the NAMCODTYP namelist block is only used to add new codes that are not already in the standard list, which is back hardcoded inside the fortran module. This means the NAMCODTYP namelist block can usually be empty.

Made several changes (many of which are already in Yves and Mike's chemistry changes coming soon) to avoid calls and code related to radiance observations when no such observations are present. An empty NAMTOV namelist block now ensures that nothing is called (assuming also that no radiance observations are provided). All other radiance namelist blocks do not need to be included at all. This will simplify future non-NWP applications of the EnVar code.