Name change from "HIMAWARI" to "HMWARI"
The changes for the geostationary satellite Himawari affect the namelist, the statsat files and also the tovs_nl_mod module. This was necessary to make the background check work for geostationary radiances. The name used by the rttov library is "HIMAWARI", but everywhere else we use the name "HMWARI".

Also, avoid calls to vtr_setup when not in analysis mode. This maintains the flexibility of using an MPI topology that is not dependent on the analysis grid dimensions when not in analysis mode.