New module ensembleStateVector_mod
This new module handles the reading of the ensemble, calculation
of the mean and access to members and other attributes of the
ensemble. It relies heavily on gridStateVector_mod. This simplifies
greatly bmatrixEnsemble_mod and will makes it easier to work with
ensembles in other applications.

Results of an analysis using ensemble covariances are changed
slightly due to a minor change in the order or removing the ensemble
mean and multiplying by a normalization factor.

Additional changes include adding the corresponding UV component
to the gridStateVector derived type when the mpi_distribution is
"VarsLevs". This facilitates horizontal interpolation and simplifies
the code significantly. When reading a file, any processor that reads
U or V will now also read the other component and store separately.