First changes to allow ensemble interpolation
Initial experiments run with something very close
to this version.

These initial changes require use of a new namelist
variable for ensemble B matrix:

EnsTopMatchesAnlTop = .true.

to indicate that the members can be safely interpolated
onto the analysis grid vertical levels since the model
top is at the same pressure for both. Will maybe try to
eliminate this variable with a check on the top
level pressure in a future commit.

Numerous changes required to allow interpolation to be
done at either single or double precision. Vertical
interpolation can also be told with a new optional
argument that the surface pressure is still in units
of hPa (as they are in standard files). Also added
ability to write to a standard file from a statevector
that is already in MPI global state. This was needed
for testing of the interpolation.