Main corrections to randompert for XC40
Updated compile script and main program to work
correctly on XC40 and PPP. Added a new namelist
variable WRITE_MPI to control whether or not to
write the perturbation files as complete files or
files that are separated over all mpi tasks.

The main oavar compile script was also split into
two parts, with all of the general "environment"
setup put in compile_commons.sh which is included
in the main compile scripts for both oavar and
randompert.

Randompert was tested on both hare and ppp2. The
mpi writing is much faster than writing complete
files on hare, but doesn't seem to help much on
ppp2, for which the cost of writing dominates.