Changes and optimization for biFourrier on the xc40
The lam spectral transform was optimized as much as possible to speed up the code on the new CRAY XC40. The transposes were modified to carry/reshuffle data as less as possible and a new set of transform using input gridded data ordered (k,i,j) was added for the optimized bens_Bsqrt formulation previously introduced by Mark.