Introduce a system-test for the LAM-biFourrier configuration
Actually, only the global mode of the oavar code is taken into account in the system-tests. A LAM confguration test named EnVar_biFourrier_XC40 was added. From now on, everybody will be responsible to make sure that their changes do not break the LAM mode as we plan to transfer a LAM-based EnVar system to CMC operations during FY 2018 for the RDPS@2.5km.

Also, the organisation of the repository was modified. All the fortran code is now under src/ whereas the maestro test suites where moved under maestro/. When compiling, a directory name compiledir/ will be created. However, the latter will be ignored by git as this directory wad added to the .gitignore file.