Reorganization of bmatrixEnsemble_mod
In order to be able to introduce new ensemble spatial covariance localization methods (e.g. gridpoint-based using a recursive-filter) in the future, a reorganization of bmatrixEnsemble_mod was performed. Essentially, all the localization-related code have been move to two new modules.

The module localization_mod is the master module calling the appropriate localization method wanted. So far, only spectral localization is possible and this code is contained in localizationspectral_mod.