Isolation of the O-P mode in a new standalone program
In order to streamlined the oavar program (main_var.ftn90), the O-P mode was be moved to a new standalone program called OminusP. This new program will be use in a new package for forecast verification.

WARNING: The results of bgck_o-p have slightly change because the trial grid is now used to defined the horizontal analysis grid.