farewell to topLevelControl_mod
topLevelControl_mod was too generic and had become a "fourre-tout". The code inside this module was either moved to other modules (e.g. OAVAR_BURP_SPLIT in burpfiles_mod.ftn90) or isolate in the routine that needs it (e.g. nconf in main_var.ftn90).

The revision number is now added directly in the main programs by recplacing the string "GIT-REVISION-NUMBER-WILL-BE-ADDED-HERE" using the sed command in the compilation scritps.