replace .ftn90 by .f90 where possible
According to Laurent C., it is recommended to used the extension .f90 whenever it is possible. This speedup the compilation, simplify the interpretation of the outputs when the execution or the compilation crash and allow a fast identification of the modules and routines using "#include", which demand a .ftn90 extension.

Therefore, all routines and modules not containing "#include" where renamed .f90