Additional check in vco setup
Now we check that the total number of levels specified
in the prototype file (usually the MM and TH records in
the analysisgrid file) equals the number of levels that
were matched to the vgrid descriptor (!!).

Also changed the TMG index number for loc_setup which
was already being used by chm obs operators - this gave
error messages in the listing.