Bug fix that affects GZ in bgck
Corrected bug in varnamelist_mod that changed
GZ from a TH variable to a MM variable without
making the corresponding changes in the obs
operator. This inconsistency caused very large
O-B values for radiosonde GZ observations
that are only treated during the background
check.

No change to main results in an assimilation
cycle, only to the background check of GZ that
will be used for verification after the next
operational delivery.

The bgck_conv system test results were updated.