Introduction of a new (optional) procedure to specify the observation central datestamp
Due to the limitation of the programs to handle the burp file format, it is not possible to specify a resume entry with another hour than 0, 6, 12 or 18 UTC. This prevent using OminusF to do hourly verification. Also, the sql to burp programs sometime write the wrong resume date in a burp file leading to the failure of OminusF.

Therefore, a new optional environment variable called OAVAR_OBS_CENTRAL_DATE was introduced to allow for an explicit specification of the datestamp in burp_setupfiles.