Initial (incomplete) changes for B-LatBands
Changes to allow calculation of new "latBands" style
of the static B matrix for global configurations.
This includes changes to calcstats for computing
the new B matrix, and addition of new module
"bMatrixLatBands_mod" for using this new matrix.

Also modified diagbmatrix so that stddev of Psi and Chi
can be output for use directly in the new B matrix
module (to allow using exactly the same stddev as with
current static B matrix).