Fix ominusf bug introduced during rebase.
Also modified oavar_system_tests suite so that
"get" task run on backend when oavar is run on
backend. This was already done for EnVar_24mem
test, but not for Var3D_splitobs and rdps tests.
So now, they are more consistent (and also work
more consistently for some reason).