Modification to fix MPI related problems in tvs_rttov_read_coefs.
In a RDPS configuration, one tile has two instruments in the BURP file
as another tiles as only one.  The code was then waiting in MPI routines
but not at the same place in the two MPI tiles causing the code to wait
forever.

Another change to bgcheck/multi_ir_bgck_mod.ftn90 (use of the original
rttov_read_coefs instead of tvs_rttov_read_coefs).