Adapt the interpolation of 'TG' in the LAM mode.
In LAM mode, the TG trial field is on another grid than other fields
like TT, UU, ...

So the program was aborting when interpolating TG on the analysis grid
when the origin grid was supposed to be the same as TT which is not the
case in LAM mode.

This commit fix that issue by taking into acount this special case.  A
mode general logic will be needed to adapt that part of the code for all
situations (see issue #46).

GitLab Fixing Issue #45