Remove ensemble interpolation from system tests.
Some system tests were interpolating themselves the ensemble members.
But, this work needs to be queued and is waiting for a long time.  Those
system tests are not to be focused on the ensemble interpolated but on
the execution of the 'VAR' task.  So we skip this interpolation step to
accelerate the execution of the tests globally.