Introduce the capability in VAR to output rehm and analysis files
A new module (addIncrement) was created to re-create what the program
addIncrement was doing. It is now possible to directly compute rebm and
anlm in the VAR step by adding WRITEANALYSIS = .TRUE. in &NAMMIN. The
default value is false.

A modification was made to vco_equal to make it more robust.