Improve ability to obtain date from input files
In the case of the program ensManip, the date is now
obtained in ens_fileName, which will use a file "pattern"
to search for a likely filename for an ensemble member.
This file is then queried for dates assuming the 2D
variable "P0" is present. For now we assume the middle
date corresponds with the center of the window. Should
really do a sort somehow on the datestamps.

Adjust 'ensManip/mean' system test to remove the date in the namelist.
Still validates.