ens_fileName: avoid problem when number of files is more than 10.
The routine 'clib_glob' return an error when the number of files matched
by the pattern is more then the maximum value given.  Here is we used 10
and in cases when we have 256 members, the pattern '*_*1' matched 25
files.  If the pattern '*_*1' returned an error, we try '*_*01' which
should be OK until we reach 1000 members. If it is still a problem, we
abort.