ens_fileName: change interface to output file base name
In this routine, we can output the file base name which as the form
'${DATE}_${ext}' which is useful when we want to use that name for
another file.  We did adjust the other calls to that routine.  For
'ens_readEnsemble', the argument 'ensFileBaseName' was no longer
needed.  For 'ens_writeEnsemble', the routine is now using
'ensFileNamePrefix' to construct file names for output.