ens_fileName: The file basename was including the full path.
This cause other calls to 'ens_filename' to add 2 times the directory of
the file.  We add logic to extract the exact basename of the file and
store it in variable 'ensFileBaseName'.

We add to modify the inputs of the 'ensManip/recenter' test because the
recentering mean must have the same basename as the ensemble members.