ensembleH program: first complete implementation
This commit includes moving subasic_obs from innovation_mod to
obsoperators_mod and now calling it automatically. Also added
optional argument to col_allocate to control if main arrays are
set to zero or not (affect memory usage). Still need to work
on efficiency of mpi communication of ensemble of columns.