ensembleH program: some efficiency improvements
To reduce communication costs, now do time interpolation
of columns before the AllToAll.

To reduce output to the listing for the nonlinear H(),
introduced an optional argument "beSilent_opt". This
was added to routines in innovation_mod, filterobs_mod,
tovs_nl_mod and obsoperators_mod.

In timcoord_mod, added an mpiglobal version of
"timeInterpWeight", needed for doing the time
interpolation before the AllToAll.