A bugfix and a modification to the LQ to HU tangent-linear transform
The bug
When only a few variables were read with many MPI tasks, the variable HUcontainsLQinFile was not set on every MPI task in gsv_readFile. Therefore, when HUcontainsLQ was set to true in gsv_readFromFile (and it was for gsv_readTrials), the HU to LQ transform was not performed correctly on every MPI task and the LQ to HU transform in variableTransforms_mod was not performed correctly
The adopted solution was simply to remove the variable HUcontainsLQinFile since it was not needed anymore. This bugfix has no impact on the unitTest results.

The modification
Until now, the reading of the trial fields in variableTransforms_mod was forcing a HU to LQ transformation. Therefore, when converting the analysis increment from LQ to HU (or vice versa), it was necessary to undo the HU to LQ transform of the trial field.
This back and forth conversion of the trial field humidity was removed. THIS CHANGES THE RESULTS where the trial HU was below the limit imposed in the HU to LQ transform (log(max(hu,MPC_MINIMUM_HU_R8))).