Loading 'rttov/12v1.1'.
This library does not contain 'lapack.o' which overrides the system
library.  This last is much faster.  EnKF experienced a factor 3
improvement in the number crunching step using this 'rttov/12v1.1'.

We do not expect to have such a spectacular improvement here in the
timings.  But, we expect some tiny changes in the results because the
eigenvalue decomposition will be affected.  This is true for tests where
we updated the results.  The test 'EnVar_96mem_ppp' is affected by a
larger margin because it impacts the IASI and SSMI/S observation
operator leading to early differences in the minimisation and since we
iterate over only 10 iterations the differences are kept somewhat larger
than the other tests.

GitLab Issue #63

We put 'memory=18G' for '/Tests/bgck_conv/UnitTest/envar/VAR' in the
system tests to satisfy the new memory settings on PPP.