Bugfix for the vertical interpolation of LAM ensembles and simplification for LAM field on a different horizontal grid
The padding routine (gsv_hPad) needed to expand the LAM ensembles frorm the coregrid to the biperiodic grid was imposing zeros for all the fields, including surface pressure. The vertical interpolation was therefore crashing for columns where Psurf = 0. gsv_hPad now impose a padding value of 1000hPa for Psurf.

Because some fields like TG are on a smaller horizontal grid than the other LAM fields, a special manipulation is needed in gsv_readFile. The previous implementaion was hardiwred to impose this extra step only when varName == TG. The new formulation does a systematic horizontal domain size check for all then variables, removing the need to do a check based on pre-defined variable names.