LQ not in gridStateVector outside B matrix modules (take 1)
A legacy of the old 3/4D-Var code is that LQ was still considered as an analysis increment variable. The conversion to ES (the observed humidity variable) was done in observation space and the output HU increment in grid point space was only obtained as a diagnostic step after the minimization.

We here restore LQ as a temporary variable (if needed) within the B matrix modules like ψ\psiψ and χ\chiχ in Bnmc.