Add suffix _opt to all optional arguments
The use of the suffix _opt standardizes how we name
optional arguments for subroutines and functions. Often
the optional argument will be associated with a local
variable of the same name, but without the suffix, which
will take the value of the argument when present and a
default value when the optional argument is not present.