Minor technical changes
These changes attempt to simplify and standardize
some three aspects of the code. The first simplifies
the passing of optional arguments to procedures that
have the same optional arguments. The second generalizes
some calls in gsv_allocate when specifying a
possibly non-standard list of variable names.

The third is the use of wrappers in utilities_mod for
all calls to ezsint/ezuvint that now include optional
arguments for setting the interpolation and extrapolation
degrees. When not specified they are explicitly set to OUR
default values of 'LINEAR' and 'MAXIMUM', respectively.
Therefore all other calls to ezsetopt have been removed.