Flexible maximum difference of altitude for surface data
Surface data were rejected based on hardwired values of maximum difference of altitude between the observation and the surface height of the model. To possibility to specify values through namelist variables was added.

For example, to set the rejection criteria to 10m for mean sea level pressure (bufcod = 10051) and 150m for surface pressure (bufcode = 10004), the following arguments must be added to the NAMFILT section.

 &NAMFILT
  nelems_altDiffMax = 2,
  list_altDiffMax(1) = 10004,
  list_altDiffMax(2) = 10051,
  value_altDiffMax(1) = 150.0,
  value_altDiffMax(2) = 10.0,

The default (legacy) values are unchanged so this commit does not impact previous results.