controlVector_mod: now setup subVectors directly in main code
By replacing the general cvm_setup with individual calls
to cvm_setupSubVector for each subVector of the total
control vector, this allows moving the setup related to
variational bias correction to be moved into the bias_setup
routine itself. As a result, the biasCorrection module could
be removed from the compilation of all programs that do
no actually need it (only used when doing an analysis).