More extensive use of new controlVector approach
These changes significantly simplify how the different
types of B matrices are accessed in bmatrix_mod. This
should make it easier when adding new types of matrices
or if we want to be able to use more than one instance
of a given B matrix type (e.g. multiple ensembles at
different resolutions). Also removed the passing of
the control vector dimension in all of the variations
of reduceToMPILocal and expandToMPIGlobal routines
related to the control vector for each type of B
matrix.

No changes to the system test results.