Change order of biasCorrection setup
The call to bias_setup needed to be moved to after
the bmatrix setup so to reproduce exactly the results
when using variational BC in the system tests.

Also renamed the module "bufr" to "bufr_mod". This was
the only module lacking this suffix.

A few wrong hardcoded "labels" were fixed in minimization
and obsSpaceDiag modules.