74: Improve tim_getDateStampFromFile
Now the file provided to tim_setup for determining
the dateStamp of the middle of the analysis window
can contain any date within the window, except the
last date. Previously, the file had to already contain
the datestamp for the middle of the window, which
can be difficult to determine for a high-level calling
routine.

While making this change, the bug in gsv_readFile was
corrected that prevented using more than 1 mpi task
for the addIncrement program. This resolves issue 76.

Also, the use of environment variables for specifying
if and how burp files are split was removed. The
use of an environment variable to specify the datstamp
was also removed, since the trial (or analysis) file
can now be used more easily for this purpose.

Related to these changes, the oMinusF program now uses
the input trial/analysis file to determine the datestamp.

The system test for addIncrement/3D was modified to use
3 time levels instead of 1. This was the initial motivation
for improving tim_getDateStampFromFile. And the number of
mpi task was increased from 1x1 to 2x2.