77: Introduce handling obs files of different formats
This modification adds a new general module for handling
observation files of different formats (types). Currently
only BURP and CMA formats are supported. Subroutines in
the new module, "obsFiles_mod", are now called instead of
directly calling those in burpFiles_mod.

All read/write routines specific to the CMA format are
now in the new module "cmaFiles_mod".

The selection of the observation file format is controlled
by the environment variable "MIDAS_OBS_FILE_TYPE" that
can be "BURP" or "CMA" for now. If not present, the module
assumes the BURP format will be used.