Fix variable 'UnitTest_run_pgm_args' in system test configurations.
When calling 'var.sh', we can specify two arguments which the second one
is the path to RAMDisk.  But, the default path is enough for these
tests.  But, we specified the second argument as 'yes' in some tests
which was wrong.

Since the RAMDisk is now used, we must adjust some resources requests
for tests 'var/EnVar/2D_ppp', 'var/Var3D/gbc' and
'var/Var3D/splitobs_ppp'.