Updates for chemical data assimilation
The following main updates are being included in
routines affecting only chemical data assimilation:

- bmatrixchem_mod.f90: Adding the capability of allowing
the static Bchm to apply to a subset of the constituent
fields of anlvar. This is to allow the input of fields
which may be required as part of observation operators
without generating increments for these fields.

- chem_obsoperators_mod.f90, chem_postproc_mod.f90,
chem_setup_mod.f90: To remove use of messaging and diagnostic
output to text files as files cannot be simultaneously
opened for writing by different processors. Depending on
timing, the possibility of this occurring is small but not
zero. As a consequence some output statements are being
commented out, removed, or modified for re-directing to the
listing file.

- obsspacediag_mod.f90: Adding to  cost function related
diagnostics for output to text file. An addition is to
distinguish between non-assimilated and assimilated O-P.
Another is to provide  A-P diagnostics.

- obssubspacedata_mo.f90: Adding of near pole obs consideration
in matching observation to auxiliary data.