LQ not in gridStateVector outside B matrix modules (take 6 and 7)
Thanks to the work of Stephen and Mark, the bug was identified: the adjoint of the LQtoHU conversion was done twice on the input statevector object because the same statevector was given to both Ben and Bhi. A simple copy of the statevector before the adjoint procedure fix the issue.

The code was then rebase with the latest version of master branch. As expected, most of the UnitTest do not validate anymore. A GDPS cycle will soon be lauched the verify if the change have an impact on the forecast quality.