Issue #108: wait4tests.sh is much less verbose
We wanted to show only once on the screen when a job aborts in the
system test maestro suite.  Before, 'wait4tests.sh' were outputting
again and again the same messages.

With this fix, we show only once each of the aborted jobs in the listing
which is more than enough.